+++ packages/salesforce-adapter/src/filters/formula_deps.ts
@@ -13,142 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 import { logger } from '@salto-io/logging'
 import { collections } from '@salto-io/lowerdash'
-import {
-  Element,
-  ElemID,
-  ElemIDType,
-  Field,
-  InstanceElement,
-  isObjectType,
-  ReadOnlyElementsSource,
-  ReferenceExpression,
-} from '@salto-io/adapter-api'
-import { extendGeneratedDependencies, naclCase } from '@salto-io/adapter-utils'
-import {
-  FormulaIdentifierInfo,
-  IdentifierType,
-  parseFormulaIdentifier,
-  extractFormulaIdentifiers,
-} from '@salto-io/salesforce-formula-parser'
+import { Element, Field, isObjectType, ReadOnlyElementsSource, ReferenceExpression } from '@salto-io/adapter-api'
+import { extendGeneratedDependencies } from '@salto-io/adapter-utils'
+import { parseFormulaIdentifier, extractFormulaIdentifiers } from '@salto-io/salesforce-formula-parser'
 import { LocalFilterCreator } from '../filter'
 import { isFormulaField } from '../transformers/transformer'
-import { CUSTOM_METADATA_SUFFIX, FORMULA, SALESFORCE } from '../constants'
-import {
-  buildElementsSourceForFetch,
-  ensureSafeFilterFetch,
-  extractFlatCustomObjectFields,
-  isInstanceOfTypeSync,
-} from './utils'
+import { FORMULA } from '../constants'
+import { buildElementsSourceForFetch, ensureSafeFilterFetch, extractFlatCustomObjectFields } from './utils'
+import { logInvalidReferences, referencesFromIdentifiers, referenceValidity } from './formula_utils'
 
 const log = logger(module)
 const { awu, groupByAsync } = collections.asynciterable
 
-const referenceFieldsWithFormulaIdentifiers: Record<string, string> = {
-  FlowCondition: 'leftValueReference',
-  FlowTestCondition: 'leftValueReference',
-  FlowTestParameter: 'leftValueReference',
-}
-
-const identifierTypeToElementName = (identifierInfo: FormulaIdentifierInfo): string[] => {
-  if (identifierInfo.type === 'customLabel') {
-    return [identifierInfo.instance]
-  }
-  if (identifierInfo.type === 'customMetadataTypeRecord') {
-    const [typeName, instanceName] = identifierInfo.instance.split('.')
-    return [`${typeName.slice(0, -1 * CUSTOM_METADATA_SUFFIX.length)}.${instanceName}`]
-  }
-  return identifierInfo.instance.split('.').slice(1)
-}
-
-const identifierTypeToElementType = (identifierInfo: FormulaIdentifierInfo): string => {
-  if (identifierInfo.type === 'customLabel') {
-    return 'CustomLabel'
-  }
-
-  return identifierInfo.instance.split('.')[0]
-}
-
-const identifierTypeToElemIdType = (identifierInfo: FormulaIdentifierInfo): ElemIDType =>
-  (
-    ({
-      standardObject: 'type',
-      customMetadataType: 'type',
-      customObject: 'type',
-      customSetting: 'type',
-      standardField: 'field',
-      customField: 'field',
-      customMetadataTypeRecord: 'instance',
-      customLabel: 'instance',
-    }) as Record<IdentifierType, ElemIDType>
-  )[identifierInfo.type]
-
-const referencesFromIdentifiers = async (typeInfos: FormulaIdentifierInfo[]): Promise<ElemID[]> =>
-  typeInfos.map(
-    identifierInfo =>
-      new ElemID(
-        SALESFORCE,
-        naclCase(identifierTypeToElementType(identifierInfo)),
-        identifierTypeToElemIdType(identifierInfo),
-        ...identifierTypeToElementName(identifierInfo).map(naclCase),
-      ),
-  )
-
-const isValidReference = async (
-  elemId: ElemID,
-  allElements: ReadOnlyElementsSource,
-): Promise<boolean> => {
-  if (elemId.idType === 'type' || elemId.idType === 'instance') {
-    return (await allElements.get(elemId)) !== undefined
-  }
-
-  // field
-  const typeElemId = new ElemID(elemId.adapter, elemId.typeName)
-  const typeElement = await allElements.get(typeElemId)
-  return (
-    typeElement !== undefined && typeElement.fields[elemId.name] !== undefined
-  )
-}
-
-const referenceValidity = async (
-  refElemId: ElemID,
-  selfElemId: ElemID,
-  allElements: ReadOnlyElementsSource,
-): Promise<'valid' | 'omitted' | 'invalid'> => {
-  const isSelfReference = (elemId: ElemID): boolean =>
-    elemId.isEqual(selfElemId)
-
-  if (isSelfReference(refElemId)) {
-    return 'omitted'
-  }
-  return (await isValidReference(refElemId, allElements)) ? 'valid' : 'invalid'
-}
-
-const logInvalidReferences = (
-  refOrigin: ElemID,
-  invalidReferences: ElemID[],
-  formula: string,
-  identifiersInfo: FormulaIdentifierInfo[][],
-): void => {
-  if (invalidReferences.length > 0) {
-    log.debug(
-      'When parsing the formula %o in %o, one or more of the identifiers %o was parsed to an invalid reference: ',
-      formula,
-      refOrigin.getFullName(),
-      identifiersInfo.flat().map((info) => info.instance),
-    )
-  }
-  invalidReferences.forEach((refElemId) => {
-    log.debug(`Invalid reference: ${refElemId.getFullName()}`)
-  })
-}
-
-const addDependenciesAnnotation = async (
-  field: Field,
-  allElements: ReadOnlyElementsSource,
-): Promise<void> => {
+const addDependenciesAnnotation = async (field: Field, allElements: ReadOnlyElementsSource): Promise<void> => {
   const formula = field.annotations[FORMULA]
   if (formula === undefined) {
     log.error(`Field ${field.elemID.getFullName()} is a formula field with no formula?`)
@@ -186,16 +65,11 @@ const addDependenciesAnnotation = async (
       References: ${references.map(ref => ref.getFullName()).join(', ')}`)
     }
 
-    const referencesWithValidity = await groupByAsync(references, (refElemId) =>
+    const referencesWithValidity = await groupByAsync(references, refElemId =>
       referenceValidity(refElemId, field.parent.elemID, allElements),
     )
 
-    logInvalidReferences(
-      field.elemID,
-      referencesWithValidity.invalid ?? [],
-      formula,
-      identifiersInfo,
-    )
+    logInvalidReferences(field.elemID, referencesWithValidity.invalid ?? [], formula, identifiersInfo)
 
     const depsAsRefExpr = (referencesWithValidity.valid ?? []).map(elemId => ({
       reference: new ReferenceExpression(elemId),
@@ -216,53 +90,7 @@ const addDependenciesToFormulaFields = async (
     .flatMap(extractFlatCustomObjectFields) // Get the types + their fields
     .filter(isFormulaField)
     .toArray()
-  await Promise.all(
-    fetchedFormulaFields.map((field) =>
-      addDependenciesAnnotation(field, allElements),
-    ),
-  )
-}
-
-const transformFieldsToReferences = async (
-  fetchedElements: Element[],
-  allElements: ReadOnlyElementsSource,
-): Promise<void> => {
-  const transformInstanceFieldToReference = async (
-    instance: InstanceElement,
-  ): Promise<void> => {
-    const fieldName =
-      referenceFieldsWithFormulaIdentifiers[instance.elemID.typeName]
-    const identifierInfo = parseFormulaIdentifier(
-      instance.value[fieldName],
-      instance.elemID.getFullName(),
-    )
-    const referenceElemIds = await referencesFromIdentifiers(identifierInfo)
-
-    const referencesWithValidity = await groupByAsync(
-      referenceElemIds,
-      (refElemId) => referenceValidity(refElemId, instance.elemID, allElements),
-    )
-
-    logInvalidReferences(
-      instance.elemID,
-      referencesWithValidity.invalid ?? [],
-      instance.value[fieldName],
-      [identifierInfo],
-    )
-
-    if (!referencesWithValidity.valid) {
-      return
-    }
-
-    instance.value[fieldName] = new ReferenceExpression(
-      referencesWithValidity.valid[0],
-    )
-  }
-
-  const fetchedInstances = fetchedElements.filter(
-    isInstanceOfTypeSync(...Object.keys(referenceFieldsWithFormulaIdentifiers)),
-  )
-  fetchedInstances.forEach(transformInstanceFieldToReference)
+  await Promise.all(fetchedFormulaFields.map(field => addDependenciesAnnotation(field, allElements)))
 }
 
 const FILTER_NAME = 'formulaDeps'
@@ -279,10 +107,9 @@ const filter: LocalFilterCreator = ({ config }) => ({
     warningMessage: 'Error while parsing formulas',
     config,
     filterName: FILTER_NAME,
-    fetchFilterFunc: async (fetchedElements) => {
+    fetchFilterFunc: async fetchedElements => {
       const allElements = buildElementsSourceForFetch(fetchedElements, config)
       await addDependenciesToFormulaFields(fetchedElements, allElements)
-      await transformFieldsToReferences(fetchedElements, allElements)
     },
   }),
 })
