+++ packages/salesforce-adapter/src/types.ts
@@ -113,6 +113,7 @@ export type OptionalFeatures = {
   describeSObjects?: boolean
   skipAliases?: boolean
   formulaDeps?: boolean
+  formulaRefFields?: boolean
   fetchCustomObjectUsingRetrieveApi?: boolean
   generateRefsInProfiles?: boolean
   fetchProfilesUsingReadApi?: boolean
@@ -807,6 +808,7 @@ const optionalFeaturesType = createMatchingObjectType<OptionalFeatures>({
     describeSObjects: { refType: BuiltinTypes.BOOLEAN },
     skipAliases: { refType: BuiltinTypes.BOOLEAN },
     formulaDeps: { refType: BuiltinTypes.BOOLEAN },
+    formulaRefFields: { refType: BuiltinTypes.BOOLEAN },
     fetchCustomObjectUsingRetrieveApi: { refType: BuiltinTypes.BOOLEAN },
     generateRefsInProfiles: { refType: BuiltinTypes.BOOLEAN },
     fetchProfilesUsingReadApi: { refType: BuiltinTypes.BOOLEAN },
