+++ packages/salesforce-adapter/test/filters/formula_deps.test.ts
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 import {
   BuiltinTypes,
   ElemID,
@@ -26,7 +25,7 @@ import { FlatDetailedDependency } from '@salto-io/adapter-utils'
 import formulaDepsFilter from '../../src/filters/formula_deps'
 import { createCustomObjectType, defaultFilterContext } from '../utils'
 import { createInstanceElement, formulaTypeName, Types } from '../../src/transformers/transformer'
-import { FIELD_TYPE_NAMES, FORMULA, METADATA_TYPE, SALESFORCE } from '../../src/constants'
+import { FIELD_TYPE_NAMES, FORMULA, SALESFORCE } from '../../src/constants'
 import { mockTypes } from '../mock_elements'
 import { FilterWith } from './mocks'
 
@@ -51,200 +50,196 @@ const customObjectTypeWithFields = (name: string, fields: Record<string, Primiti
 describe('Formula dependencies', () => {
   let filter: FilterWith<'onFetch'>
 
+  let typeWithFormula: ObjectType
+  let referredTypes: ObjectType[]
+  let referredInstances: InstanceElement[]
+
   beforeAll(() => {
     const config = { ...defaultFilterContext }
     filter = formulaDepsFilter({ config }) as FilterWith<'onFetch'>
   })
 
-  describe('Formula fields', () => {
-    let typeWithFormula: ObjectType
-    let referredTypes: ObjectType[]
-    let referredInstances: InstanceElement[]
-
-    beforeAll(() => {
-      typeWithFormula = createCustomObjectType('Account', {
-        fields: {
-          someField__c: {
-            refType: BuiltinTypes.STRING,
-          },
-          someFormulaField__c: {
-            refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.CHECKBOX)],
-            annotations: {
-              [FORMULA]: 'ISBLANK(someField__c)',
-            },
-          },
-          AccountNumber: {
-            refType: BuiltinTypes.NUMBER,
-          },
-          OwnerId: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          original_lead__c: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          LastModifiedById: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          OpportunityId: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          Opportunity__c: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          ParentId: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
-          },
-          RecordTypeId: {
-            refType: BuiltinTypes.SERVICE_ID_NUMBER,
+  beforeAll(() => {
+    typeWithFormula = createCustomObjectType('Account', {
+      fields: {
+        someField__c: {
+          refType: BuiltinTypes.STRING,
+        },
+        someFormulaField__c: {
+          refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.CHECKBOX)],
+          annotations: {
+            [FORMULA]: 'ISBLANK(someField__c)',
           },
         },
-      })
-      const someCustomMetadataType = customObjectTypeWithFields('Trigger_Context_Status__mdt', {
-        Enable_After_Delete__c: BuiltinTypes.BOOLEAN,
-        Enable_After_Insert__c: BuiltinTypes.BOOLEAN,
-        DeveloperName: BuiltinTypes.STRING,
-      })
-
-      referredTypes = [
-        customObjectTypeWithFields('Contact', {
-          AccountId: BuiltinTypes.SERVICE_ID_NUMBER,
-          AssistantName: BuiltinTypes.STRING,
-          CreatedById: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        someCustomMetadataType,
-        customObjectTypeWithFields('User', {
-          ContactId: BuiltinTypes.SERVICE_ID_NUMBER,
-          ManagerId: BuiltinTypes.SERVICE_ID_NUMBER,
-          CompanyName: BuiltinTypes.STRING,
-          ProfileId: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('original_lead__r', {
-          ConvertedAccountId: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('Center__c', {
-          My_text_field__c: BuiltinTypes.STRING,
-        }),
-        customObjectTypeWithFields('Customer_Support_Setting__c', {
-          Email_Address__c: BuiltinTypes.STRING,
-        }),
-        createCustomObjectType('Details', {}),
-        customObjectTypeWithFields('Opportunity', {
-          AccountId: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('Opportunity__r', {
-          Related_Asset__c: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('Organization', {
-          UiSkin: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('Profile', {
-          Id: BuiltinTypes.SERVICE_ID_NUMBER,
-        }),
-        customObjectTypeWithFields('RecordType', { Name: BuiltinTypes.STRING }),
-        customObjectTypeWithFields('Related_Asset__r', {
-          Name: BuiltinTypes.STRING,
-        }),
-        customObjectTypeWithFields('SRM_API_Metadata_Client_Setting__mdt', {
-          CreatedDate: BuiltinTypes.STRING,
-        }),
-      ]
-      referredInstances = [
-        createInstanceElement({ fullName: 'Details' }, mockTypes.CustomLabel),
-        createInstanceElement({ fullName: 'Trigger_Context_Status.by_class' }, someCustomMetadataType),
-
-        createInstanceElement({ fullName: 'Trigger_Context_Status.by_handler' }, someCustomMetadataType),
-      ]
+        AccountNumber: {
+          refType: BuiltinTypes.NUMBER,
+        },
+        OwnerId: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        original_lead__c: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        LastModifiedById: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        OpportunityId: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        Opportunity__c: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        ParentId: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+        RecordTypeId: {
+          refType: BuiltinTypes.SERVICE_ID_NUMBER,
+        },
+      },
+    })
+    const someCustomMetadataType = customObjectTypeWithFields('Trigger_Context_Status__mdt', {
+      Enable_After_Delete__c: BuiltinTypes.BOOLEAN,
+      Enable_After_Insert__c: BuiltinTypes.BOOLEAN,
+      DeveloperName: BuiltinTypes.STRING,
     })
 
-    describe('When the formula has a reference', () => {
-      it('Should return a field reference as a dependency', async () => {
-        const elements = [typeWithFormula.clone()]
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          elements[0].fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(deps[0]).toEqual(depNameToRefExpr(typeWithFormula.elemID.typeName, 'someField__c'))
+    referredTypes = [
+      customObjectTypeWithFields('Contact', {
+        AccountId: BuiltinTypes.SERVICE_ID_NUMBER,
+        AssistantName: BuiltinTypes.STRING,
+        CreatedById: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      someCustomMetadataType,
+      customObjectTypeWithFields('User', {
+        ContactId: BuiltinTypes.SERVICE_ID_NUMBER,
+        ManagerId: BuiltinTypes.SERVICE_ID_NUMBER,
+        CompanyName: BuiltinTypes.STRING,
+        ProfileId: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('original_lead__r', {
+        ConvertedAccountId: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('Center__c', {
+        My_text_field__c: BuiltinTypes.STRING,
+      }),
+      customObjectTypeWithFields('Customer_Support_Setting__c', {
+        Email_Address__c: BuiltinTypes.STRING,
+      }),
+      createCustomObjectType('Details', {}),
+      customObjectTypeWithFields('Opportunity', {
+        AccountId: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('Opportunity__r', {
+        Related_Asset__c: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('Organization', {
+        UiSkin: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('Profile', {
+        Id: BuiltinTypes.SERVICE_ID_NUMBER,
+      }),
+      customObjectTypeWithFields('RecordType', { Name: BuiltinTypes.STRING }),
+      customObjectTypeWithFields('Related_Asset__r', {
+        Name: BuiltinTypes.STRING,
+      }),
+      customObjectTypeWithFields('SRM_API_Metadata_Client_Setting__mdt', {
+        CreatedDate: BuiltinTypes.STRING,
+      }),
+    ]
+    referredInstances = [
+      createInstanceElement({ fullName: 'Details' }, mockTypes.CustomLabel),
+      createInstanceElement({ fullName: 'Trigger_Context_Status.by_class' }, someCustomMetadataType),
+
+      createInstanceElement({ fullName: 'Trigger_Context_Status.by_handler' }, someCustomMetadataType),
+    ]
+  })
+
+  describe('When the formula has a reference', () => {
+    it('Should return a field reference as a dependency', async () => {
+      const elements = [typeWithFormula.clone()]
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        elements[0].fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(deps[0]).toEqual(depNameToRefExpr(typeWithFormula.elemID.typeName, 'someField__c'))
+    })
+    it('Should return a Metadata Type Record field reference as a dependency', async () => {
+      const customMetadataType = customObjectTypeWithFields('SomeCustomMetadataType__mdt', {
+        SomeTextField__c: BuiltinTypes.STRING,
       })
-      it('Should return a Metadata Type Record field reference as a dependency', async () => {
-        const customMetadataType = customObjectTypeWithFields('SomeCustomMetadataType__mdt', {
-          SomeTextField__c: BuiltinTypes.STRING,
-        })
-        const typeUnderTest = typeWithFormula.clone()
-        typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
-          '$CustomMetadata.SomeCustomMetadataType__mdt.SomeCustomMetadataTypeRecord.SomeTextField__c'
-        const elements = [
-          typeUnderTest,
-          customMetadataType,
-          createInstanceElement(
-            { fullName: 'SomeCustomMetadataType.SomeCustomMetadataTypeRecord' },
-            customMetadataType,
+      const typeUnderTest = typeWithFormula.clone()
+      typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
+        '$CustomMetadata.SomeCustomMetadataType__mdt.SomeCustomMetadataTypeRecord.SomeTextField__c'
+      const elements = [
+        typeUnderTest,
+        customMetadataType,
+        createInstanceElement({ fullName: 'SomeCustomMetadataType.SomeCustomMetadataTypeRecord' }, customMetadataType),
+      ]
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+
+      const expectedRefs = [
+        {
+          reference: new ReferenceExpression(new ElemID(SALESFORCE, 'SomeCustomMetadataType__mdt')),
+        },
+        {
+          reference: new ReferenceExpression(
+            new ElemID(SALESFORCE, 'SomeCustomMetadataType__mdt', 'field', 'SomeTextField__c'),
           ),
-        ]
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-
-        const expectedRefs = [
-          {
-            reference: new ReferenceExpression(new ElemID(SALESFORCE, 'SomeCustomMetadataType__mdt')),
-          },
-          {
-            reference: new ReferenceExpression(
-              new ElemID(SALESFORCE, 'SomeCustomMetadataType__mdt', 'field', 'SomeTextField__c'),
+        },
+        {
+          reference: new ReferenceExpression(
+            new ElemID(
+              SALESFORCE,
+              'SomeCustomMetadataType__mdt',
+              'instance',
+              'SomeCustomMetadataType_SomeCustomMetadataTypeRecord@v',
             ),
-          },
-          {
-            reference: new ReferenceExpression(
-              new ElemID(
-                SALESFORCE,
-                'SomeCustomMetadataType__mdt',
-                'instance',
-                'SomeCustomMetadataType_SomeCustomMetadataTypeRecord@v',
-              ),
-            ),
-          },
-        ]
-        expect(deps).toEqual(expectedRefs)
-      })
-      it("should not return a reference if it refers to something that doesn't exist", async () => {
-        const typeUnderTest = typeWithFormula.clone()
-        typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
-          'ISBLANK(someField__c) && ISBLANK(GarbageType.DreckField)'
-
-        await filter.onFetch([typeUnderTest])
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(deps).toHaveLength(1)
-        expect(deps[0].reference.elemID.typeName).not.toEqual('GarbageType')
-      })
+          ),
+        },
+      ]
+      expect(deps).toEqual(expectedRefs)
     })
+    it("should not return a reference if it refers to something that doesn't exist", async () => {
+      const typeUnderTest = typeWithFormula.clone()
+      typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
+        'ISBLANK(someField__c) && ISBLANK(GarbageType.DreckField)'
 
-    describe('When referencing RecordType', () => {
-      it('should extract the correct dependencies', async () => {
-        const typeUnderTest = typeWithFormula.clone()
-        const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
-        typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
-          `IF( RecordType.Name = 'New Sale', 'Onboarding - New',
+      await filter.onFetch([typeUnderTest])
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(deps).toHaveLength(1)
+      expect(deps[0].reference.elemID.typeName).not.toEqual('GarbageType')
+    })
+  })
+
+  describe('When referencing RecordType', () => {
+    it('should extract the correct dependencies', async () => {
+      const typeUnderTest = typeWithFormula.clone()
+      const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
+      typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] =
+        `IF( RecordType.Name = 'New Sale', 'Onboarding - New',
           IF( RecordType.Name = 'Upgrade', 'Onboarding - Upgrade',
             IF( RecordType.Name = 'Add-On', 'Onboarding - Add-On', "")))`
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(deps[0]).toEqual(depNameToRefExpr(typeWithFormula.elemID.typeName, 'RecordTypeId'))
-        expect(deps[1]).toEqual(depNameToRefExpr('RecordType'))
-        expect(deps[2]).toEqual(depNameToRefExpr('RecordType', 'Name'))
-      })
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(deps[0]).toEqual(depNameToRefExpr(typeWithFormula.elemID.typeName, 'RecordTypeId'))
+      expect(deps[1]).toEqual(depNameToRefExpr('RecordType'))
+      expect(deps[2]).toEqual(depNameToRefExpr('RecordType', 'Name'))
     })
+  })
 
-    describe('Complex formulas', () => {
-      const standardFormula = `IF(Owner.Contact.CreatedBy.Manager.Profile.Id = "03d3h000000khEQ",TRUE,false)
+  describe('Complex formulas', () => {
+    const standardFormula = `IF(Owner.Contact.CreatedBy.Manager.Profile.Id = "03d3h000000khEQ",TRUE,false)
       && IF(($CustomMetadata.Trigger_Context_Status__mdt.by_handler.Enable_After_Insert__c ||
         $CustomMetadata.Trigger_Context_Status__mdt.by_class.DeveloperName = "Default"),true,FALSE)
       && IF( ($Label.Details = "Value" || Opportunity.Account.Parent.Parent.Parent.LastModifiedBy.Contact.AssistantName = "Marie"), true ,false)
@@ -254,161 +249,161 @@ describe('Formula dependencies', () => {
       && IF ( ( TEXT($Organization.UiSkin) = "lex" ) ,true,false)    
       && IF ( ( $Setup.Customer_Support_Setting__c.Email_Address__c = "test@gmail.com" ) ,true,false)    
       && IF ( (  $User.CompanyName = "acme" ) ,true,false)`
-      const processBuilderFormula = `IF([Account].Owner.Manager.Contact.Account.AccountNumber  = "text" ,TRUE,FALSE)
+    const processBuilderFormula = `IF([Account].Owner.Manager.Contact.Account.AccountNumber  = "text" ,TRUE,FALSE)
           || IF([Account].original_lead__r.ConvertedAccountId != "",TRUE,FALSE)
           || IF($CustomMetadata.Trigger_Context_Status__mdt.by_class.Enable_After_Delete__c , TRUE,FALse)`
-      const standardFormulaExpectedRefs = [
-        'salesforce.Account.field.LastModifiedById',
-        'salesforce.Account.field.OpportunityId',
-        'salesforce.Account.field.Opportunity__c',
-        'salesforce.Account.field.OwnerId',
-        'salesforce.Account.field.ParentId',
-        'salesforce.Center__c',
-        'salesforce.Center__c.field.My_text_field__c',
-        'salesforce.Contact',
-        'salesforce.Contact.field.AssistantName',
-        'salesforce.Contact.field.CreatedById',
-        'salesforce.CustomLabel.instance.Details',
-        'salesforce.Customer_Support_Setting__c',
-        'salesforce.Customer_Support_Setting__c.field.Email_Address__c',
-        'salesforce.Opportunity',
-        'salesforce.Opportunity.field.AccountId',
-        'salesforce.Opportunity__r',
-        'salesforce.Opportunity__r.field.Related_Asset__c',
-        'salesforce.Organization',
-        'salesforce.Organization.field.UiSkin',
-        'salesforce.Profile',
-        'salesforce.Profile.field.Id',
-        'salesforce.Related_Asset__r',
-        'salesforce.Related_Asset__r.field.Name',
-        'salesforce.SRM_API_Metadata_Client_Setting__mdt',
-        'salesforce.SRM_API_Metadata_Client_Setting__mdt.field.CreatedDate',
-        'salesforce.Trigger_Context_Status__mdt',
-        'salesforce.Trigger_Context_Status__mdt.field.DeveloperName',
-        'salesforce.Trigger_Context_Status__mdt.field.Enable_After_Insert__c',
-        'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_class@uuvu',
-        'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_handler@uuvu',
-        'salesforce.User',
-        'salesforce.User.field.CompanyName',
-        'salesforce.User.field.ContactId',
-        'salesforce.User.field.ManagerId',
-        'salesforce.User.field.ProfileId',
-      ]
-      const processBuilderFormulaExpectedRefs = [
-        'salesforce.Account.field.AccountNumber',
-        'salesforce.Account.field.OwnerId',
-        'salesforce.Account.field.original_lead__c',
-        'salesforce.Contact',
-        'salesforce.Contact.field.AccountId',
-        'salesforce.Trigger_Context_Status__mdt',
-        'salesforce.Trigger_Context_Status__mdt.field.Enable_After_Delete__c',
-        'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_class@uuvu',
-        'salesforce.User',
-        'salesforce.User.field.ContactId',
-        'salesforce.User.field.ManagerId',
-        'salesforce.original_lead__r',
-        'salesforce.original_lead__r.field.ConvertedAccountId',
-      ]
-      it('Should extract the correct references from a complex standard formula', async () => {
-        const typeUnderTest = typeWithFormula.clone()
-        typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] = standardFormula
-        const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(deps.map(({ reference }: { reference: ReferenceExpression }) => reference.elemID.getFullName())).toEqual(
-          standardFormulaExpectedRefs,
-        )
-      })
-
-      it('Should extract the correct references from a Process Builder formula', async () => {
-        const typeUnderTest = typeWithFormula.clone()
-        typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] = processBuilderFormula
-        const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(
-          deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()),
-        ).toEqual(processBuilderFormulaExpectedRefs)
-      })
+    const standardFormulaExpectedRefs = [
+      'salesforce.Account.field.LastModifiedById',
+      'salesforce.Account.field.OpportunityId',
+      'salesforce.Account.field.Opportunity__c',
+      'salesforce.Account.field.OwnerId',
+      'salesforce.Account.field.ParentId',
+      'salesforce.Center__c',
+      'salesforce.Center__c.field.My_text_field__c',
+      'salesforce.Contact',
+      'salesforce.Contact.field.AssistantName',
+      'salesforce.Contact.field.CreatedById',
+      'salesforce.CustomLabel.instance.Details',
+      'salesforce.Customer_Support_Setting__c',
+      'salesforce.Customer_Support_Setting__c.field.Email_Address__c',
+      'salesforce.Opportunity',
+      'salesforce.Opportunity.field.AccountId',
+      'salesforce.Opportunity__r',
+      'salesforce.Opportunity__r.field.Related_Asset__c',
+      'salesforce.Organization',
+      'salesforce.Organization.field.UiSkin',
+      'salesforce.Profile',
+      'salesforce.Profile.field.Id',
+      'salesforce.Related_Asset__r',
+      'salesforce.Related_Asset__r.field.Name',
+      'salesforce.SRM_API_Metadata_Client_Setting__mdt',
+      'salesforce.SRM_API_Metadata_Client_Setting__mdt.field.CreatedDate',
+      'salesforce.Trigger_Context_Status__mdt',
+      'salesforce.Trigger_Context_Status__mdt.field.DeveloperName',
+      'salesforce.Trigger_Context_Status__mdt.field.Enable_After_Insert__c',
+      'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_class@uuvu',
+      'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_handler@uuvu',
+      'salesforce.User',
+      'salesforce.User.field.CompanyName',
+      'salesforce.User.field.ContactId',
+      'salesforce.User.field.ManagerId',
+      'salesforce.User.field.ProfileId',
+    ]
+    const processBuilderFormulaExpectedRefs = [
+      'salesforce.Account.field.AccountNumber',
+      'salesforce.Account.field.OwnerId',
+      'salesforce.Account.field.original_lead__c',
+      'salesforce.Contact',
+      'salesforce.Contact.field.AccountId',
+      'salesforce.Trigger_Context_Status__mdt',
+      'salesforce.Trigger_Context_Status__mdt.field.Enable_After_Delete__c',
+      'salesforce.Trigger_Context_Status__mdt.instance.Trigger_Context_Status_by_class@uuvu',
+      'salesforce.User',
+      'salesforce.User.field.ContactId',
+      'salesforce.User.field.ManagerId',
+      'salesforce.original_lead__r',
+      'salesforce.original_lead__r.field.ConvertedAccountId',
+    ]
+    it('Should extract the correct references from a complex standard formula', async () => {
+      const typeUnderTest = typeWithFormula.clone()
+      typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] = standardFormula
+      const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(deps.map(({ reference }: { reference: ReferenceExpression }) => reference.elemID.getFullName())).toEqual(
+        standardFormulaExpectedRefs,
+      )
     })
-    describe('CPQ formulas', () => {
-      const simpleFormula = 'SBQQ__DistriBUtor__r.Name '
-      const formulaWithUnknownRelationship = 'SBQQ__random__r.Name '
-      const simpleFormulaExpectedDeps = [
-        'salesforce.SBQQ__QuoTE__c.field.SBQQ__DistriBUtor__c',
-        'salesforce.Account',
-        'salesforce.Account.field.Name',
-      ].sort()
-      const formulaWithUnknownRelationshipExpectedDeps = [
-        'salesforce.SBQQ__QuoTE__c.field.SBQQ__random__c',
-        'salesforce.SBQQ__random__r',
-        'salesforce.SBQQ__random__r.field.Name',
-      ].sort()
-      let typeWithCpqFormula: ObjectType
-      let referredObjectTypes: ObjectType[]
 
-      beforeAll(() => {
-        typeWithCpqFormula = createCustomObjectType('SBQQ__QuoTE__c', {
-          fields: {
-            SBQQ__DistriBUtor__c: {
-              refType: BuiltinTypes.STRING,
-            },
-            SBQQ__random__c: {
-              refType: BuiltinTypes.STRING,
-            },
-            someField__c: {
-              refType: BuiltinTypes.STRING,
-            },
-            someFormulaField__c: {
-              refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.CHECKBOX)],
-              annotations: {
-                [FORMULA]: '',
-              },
+    it('Should extract the correct references from a Process Builder formula', async () => {
+      const typeUnderTest = typeWithFormula.clone()
+      typeUnderTest.fields.someFormulaField__c.annotations[FORMULA] = processBuilderFormula
+      const elements = [typeUnderTest, ...referredTypes, ...referredInstances]
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        typeUnderTest.fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName())).toEqual(
+        processBuilderFormulaExpectedRefs,
+      )
+    })
+  })
+  describe('CPQ formulas', () => {
+    const simpleFormula = 'SBQQ__DistriBUtor__r.Name '
+    const formulaWithUnknownRelationship = 'SBQQ__random__r.Name '
+    const simpleFormulaExpectedDeps = [
+      'salesforce.SBQQ__QuoTE__c.field.SBQQ__DistriBUtor__c',
+      'salesforce.Account',
+      'salesforce.Account.field.Name',
+    ].sort()
+    const formulaWithUnknownRelationshipExpectedDeps = [
+      'salesforce.SBQQ__QuoTE__c.field.SBQQ__random__c',
+      'salesforce.SBQQ__random__r',
+      'salesforce.SBQQ__random__r.field.Name',
+    ].sort()
+    let typeWithCpqFormula: ObjectType
+    let referredObjectTypes: ObjectType[]
+
+    beforeAll(() => {
+      typeWithCpqFormula = createCustomObjectType('SBQQ__QuoTE__c', {
+        fields: {
+          SBQQ__DistriBUtor__c: {
+            refType: BuiltinTypes.STRING,
+          },
+          SBQQ__random__c: {
+            refType: BuiltinTypes.STRING,
+          },
+          someField__c: {
+            refType: BuiltinTypes.STRING,
+          },
+          someFormulaField__c: {
+            refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.CHECKBOX)],
+            annotations: {
+              [FORMULA]: '',
             },
           },
-        })
-        referredObjectTypes = [
-          customObjectTypeWithFields('SBQQ__random__r', {
-            Name: BuiltinTypes.STRING,
-          }),
-          customObjectTypeWithFields('Account', { Name: BuiltinTypes.STRING }),
-        ]
+        },
       })
-
-      it('Should extract the correct references from a CPQ formula', async () => {
-        const elements = [typeWithCpqFormula.clone(), ...referredObjectTypes]
-        elements[0].fields.someFormulaField__c.annotations[FORMULA] = simpleFormula
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          elements[0].fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(
-          deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
-        ).toEqual(simpleFormulaExpectedDeps)
-      })
-
-      it('Should extract the correct references from a CPQ formula with unknown relationship', async () => {
-        const elements = [typeWithCpqFormula.clone(), ...referredObjectTypes]
-        elements[0].fields.someFormulaField__c.annotations[FORMULA] = formulaWithUnknownRelationship
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          elements[0].fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(
-          deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
-        ).toEqual(formulaWithUnknownRelationshipExpectedDeps)
-      })
+      referredObjectTypes = [
+        customObjectTypeWithFields('SBQQ__random__r', {
+          Name: BuiltinTypes.STRING,
+        }),
+        customObjectTypeWithFields('Account', { Name: BuiltinTypes.STRING }),
+      ]
     })
-    describe('Deeply nested formulas', () => {
-      const deeplyNestedFormula = `
+
+    it('Should extract the correct references from a CPQ formula', async () => {
+      const elements = [typeWithCpqFormula.clone(), ...referredObjectTypes]
+      elements[0].fields.someFormulaField__c.annotations[FORMULA] = simpleFormula
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        elements[0].fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(
+        deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
+      ).toEqual(simpleFormulaExpectedDeps)
+    })
+
+    it('Should extract the correct references from a CPQ formula with unknown relationship', async () => {
+      const elements = [typeWithCpqFormula.clone(), ...referredObjectTypes]
+      elements[0].fields.someFormulaField__c.annotations[FORMULA] = formulaWithUnknownRelationship
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        elements[0].fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(
+        deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
+      ).toEqual(formulaWithUnknownRelationshipExpectedDeps)
+    })
+  })
+  describe('Deeply nested formulas', () => {
+    const deeplyNestedFormula = `
       IF(INCLUDES( COVID19_Status__c  ,'Fraud evaluation on hold'),'Negative',
         IF(INCLUDES(COVID19_Status__c,'Volume increase'),'Positive',
           IF(INCLUDES(COVID19_Status__c,'Volume decrease'),'Negative',
@@ -420,79 +415,31 @@ describe('Formula dependencies', () => {
                       IF(INCLUDES(COVID19_Status__c,'Paused eCommerce operations'),'Negative',
                         IF(INCLUDES(COVID19_Status__c,'Started ecomm operations'),'Positive',""))))))))))
       `
-      const typeWithDeeplyNestedFormula = createCustomObjectType('Account', {
-        fields: {
-          COVID19_Status__c: {
-            refType: BuiltinTypes.STRING,
-          },
-          someFormulaField__c: {
-            refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.TEXT)],
-            annotations: {
-              [FORMULA]: '',
-            },
-          },
-        },
-      })
-
-      it('should not take too long', async () => {
-        const elements = [typeWithDeeplyNestedFormula.clone()]
-        elements[0].fields.someFormulaField__c.annotations[FORMULA] = deeplyNestedFormula
-        await filter.onFetch(elements)
-        const deps =
-          // eslint-disable-next-line no-underscore-dangle
-          elements[0].fields.someFormulaField__c.annotations._generated_dependencies
-        expect(deps).toBeDefined()
-        expect(
-          deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
-        ).toEqual(['salesforce.Account.field.COVID19_Status__c'])
-      })
-    })
-  })
-
-  describe('Reference fields', () => {
-    const objectType = new ObjectType({
-      elemID: new ElemID(SALESFORCE, 'FlowCondition'),
-      annotations: { [METADATA_TYPE]: 'FlowCondition' },
+    const typeWithDeeplyNestedFormula = createCustomObjectType('Account', {
       fields: {
-        leftValueReference: {
+        COVID19_Status__c: {
           refType: BuiltinTypes.STRING,
+        },
+        someFormulaField__c: {
+          refType: Types.formulaDataTypes[formulaTypeName(FIELD_TYPE_NAMES.TEXT)],
+          annotations: {
+            [FORMULA]: '',
+          },
         },
       },
     })
-    const referringInstance = new InstanceElement('SomeFlowCondition', objectType, {
-      leftValueReference: '$Label.SomeLabel',
-    })
 
-    describe('when there is a valid reference', () => {
-      const referredInstance = new InstanceElement('SomeLabel', mockTypes.CustomLabel, { fullName: 'SomeLabel' })
-      const elements = [objectType, referringInstance, referredInstance].map(element => element.clone())
-
-      beforeEach(async () => {
-        await filter.onFetch(elements)
-      })
-
-      it('should replace the field value with a reference expression', () => {
-        const referringInstanceAfterTest = elements.find(elem =>
-          elem.elemID.isEqual(referringInstance.elemID),
-        ) as InstanceElement
-        expect(referringInstanceAfterTest.value.leftValueReference).toEqual(
-          new ReferenceExpression(referredInstance.elemID),
-        )
-      })
-    })
-    describe('when the reference is not valid', () => {
-      const elements = [objectType, referringInstance].map(element => element.clone())
-
-      beforeEach(async () => {
-        await filter.onFetch(elements)
-      })
-
-      it('should replace the field value with a reference expression', () => {
-        const referringInstanceAfterTest = elements.find(elem =>
-          elem.elemID.isEqual(referringInstance.elemID),
-        ) as InstanceElement
-        expect(referringInstanceAfterTest.value.leftValueReference).toEqual(referringInstance.value.leftValueReference)
-      })
+    it('should not take too long', async () => {
+      const elements = [typeWithDeeplyNestedFormula.clone()]
+      elements[0].fields.someFormulaField__c.annotations[FORMULA] = deeplyNestedFormula
+      await filter.onFetch(elements)
+      const deps =
+        // eslint-disable-next-line no-underscore-dangle
+        elements[0].fields.someFormulaField__c.annotations._generated_dependencies
+      expect(deps).toBeDefined()
+      expect(
+        deps.map((refExpr: { reference: ReferenceExpression }) => refExpr.reference.elemID.getFullName()).sort(),
+      ).toEqual(['salesforce.Account.field.COVID19_Status__c'])
     })
   })
 })
